requirements.txt
botbuilder-integration-aiohttp>=4.10.1
botbuilder-core
botbuilder-ai
flask
